src/langchain/requirements.txt
langchain
langchain-community
langgraph
langgraph-checkpoint-sqlite
python-dotenv
pymongo
pandas
openai
llama-cpp-python